fix: text alignement in cell renderers
Dates (in any format) should ideally be left-aligned inside a data table column.

A rule of thumb is: integers should be right aligned, while other types of data should be left aligned.

The default alignment for most types of data is left alignment -- this helps make the data easily scannable, readable and comparable.

https://ux.stackexchange.com/questions/117685/in-tables-should-date-time-year-in-table-be-left-or-right-aligned#:~:text=Dates%20(in%20any%20format)%20should,data%20should%20be%20left%20aligned.